--- a/nntrainer/tensor/hgemm/hgemm_kernel_4x4.h
+++ b/nntrainer/tensor/hgemm/hgemm_kernel_4x4.h
@@ -14,7 +14,115 @@
 #include <hgemm_common.h>
 #include <stdlib.h>
 
-// 1. Partial sum 128 digits : worst accuracy, best latency
+// 1. Partial sum 256 digits : Medium accuracy, medium latency
+#define KERNEL_4x4_ACC16()               \
+  v24 = vdup_n_f16(0.F);                 \
+  v25 = vdup_n_f16(0.F);                 \
+  v26 = vdup_n_f16(0.F);                 \
+  v27 = vdup_n_f16(0.F);                 \
+  dv0 = vld1_f16(a);                     \
+  vb0 = vld1_f16(b);                     \
+  v24 = vfma_lane_f16(v24, vb0, dv0, 0); \
+  v25 = vfma_lane_f16(v25, vb0, dv0, 1); \
+  v26 = vfma_lane_f16(v26, vb0, dv0, 2); \
+  v27 = vfma_lane_f16(v27, vb0, dv0, 3); \
+  dv1 = vld1_f16(a + 4);                 \
+  vb1 = vld1_f16(b + 4);                 \
+  v24 = vfma_lane_f16(v24, vb1, dv1, 0); \
+  v25 = vfma_lane_f16(v25, vb1, dv1, 1); \
+  v26 = vfma_lane_f16(v26, vb1, dv1, 2); \
+  v27 = vfma_lane_f16(v27, vb1, dv1, 3); \
+  dv2 = vld1_f16(a + 4 * 2);             \
+  vb2 = vld1_f16(b + 4 * 2);             \
+  v24 = vfma_lane_f16(v24, vb2, dv2, 0); \
+  v25 = vfma_lane_f16(v25, vb2, dv2, 1); \
+  v26 = vfma_lane_f16(v26, vb2, dv2, 2); \
+  v27 = vfma_lane_f16(v27, vb2, dv2, 3); \
+  dv3 = vld1_f16(a + 4 * 3);             \
+  vb3 = vld1_f16(b + 4 * 3);             \
+  v24 = vfma_lane_f16(v24, vb3, dv3, 0); \
+  v25 = vfma_lane_f16(v25, vb3, dv3, 1); \
+  v26 = vfma_lane_f16(v26, vb3, dv3, 2); \
+  v27 = vfma_lane_f16(v27, vb3, dv3, 3); \
+  dv4 = vld1_f16(a + 4 * 4);             \
+  vb4 = vld1_f16(b + 4 * 4);             \
+  v24 = vfma_lane_f16(v24, vb4, dv4, 0); \
+  v25 = vfma_lane_f16(v25, vb4, dv4, 1); \
+  v26 = vfma_lane_f16(v26, vb4, dv4, 2); \
+  v27 = vfma_lane_f16(v27, vb4, dv4, 3); \
+  dv5 = vld1_f16(a + 4 * 5);             \
+  vb5 = vld1_f16(b + 4 * 5);             \
+  v24 = vfma_lane_f16(v24, vb5, dv5, 0); \
+  v25 = vfma_lane_f16(v25, vb5, dv5, 1); \
+  v26 = vfma_lane_f16(v26, vb5, dv5, 2); \
+  v27 = vfma_lane_f16(v27, vb5, dv5, 3); \
+  dv6 = vld1_f16(a + 4 * 6);             \
+  vb6 = vld1_f16(b + 4 * 6);             \
+  v24 = vfma_lane_f16(v24, vb6, dv6, 0); \
+  v25 = vfma_lane_f16(v25, vb6, dv6, 1); \
+  v26 = vfma_lane_f16(v26, vb6, dv6, 2); \
+  v27 = vfma_lane_f16(v27, vb6, dv6, 3); \
+  dv7 = vld1_f16(a + 4 * 7);             \
+  vb7 = vld1_f16(b + 4 * 7);             \
+  v24 = vfma_lane_f16(v24, vb7, dv7, 0); \
+  v25 = vfma_lane_f16(v25, vb7, dv7, 1); \
+  v26 = vfma_lane_f16(v26, vb7, dv7, 2); \
+  v27 = vfma_lane_f16(v27, vb7, dv7, 3); \
+  dv7 = vld1_f16(a + 4 * 8);             \
+  vb7 = vld1_f16(b + 4 * 8);             \
+  v24 = vfma_lane_f16(v24, vb7, dv7, 0); \
+  v25 = vfma_lane_f16(v25, vb7, dv7, 1); \
+  v26 = vfma_lane_f16(v26, vb7, dv7, 2); \
+  v27 = vfma_lane_f16(v27, vb7, dv7, 3); \
+  dv7 = vld1_f16(a + 4 * 9);             \
+  vb7 = vld1_f16(b + 4 * 9);             \
+  v24 = vfma_lane_f16(v24, vb7, dv7, 0); \
+  v25 = vfma_lane_f16(v25, vb7, dv7, 1); \
+  v26 = vfma_lane_f16(v26, vb7, dv7, 2); \
+  v27 = vfma_lane_f16(v27, vb7, dv7, 3); \
+  dv7 = vld1_f16(a + 4 * 10);            \
+  vb7 = vld1_f16(b + 4 * 10);            \
+  v24 = vfma_lane_f16(v24, vb7, dv7, 0); \
+  v25 = vfma_lane_f16(v25, vb7, dv7, 1); \
+  v26 = vfma_lane_f16(v26, vb7, dv7, 2); \
+  v27 = vfma_lane_f16(v27, vb7, dv7, 3); \
+  dv7 = vld1_f16(a + 4 * 11);            \
+  vb7 = vld1_f16(b + 4 * 11);            \
+  v24 = vfma_lane_f16(v24, vb7, dv7, 0); \
+  v25 = vfma_lane_f16(v25, vb7, dv7, 1); \
+  v26 = vfma_lane_f16(v26, vb7, dv7, 2); \
+  v27 = vfma_lane_f16(v27, vb7, dv7, 3); \
+  dv7 = vld1_f16(a + 4 * 12);            \
+  vb7 = vld1_f16(b + 4 * 12);            \
+  v24 = vfma_lane_f16(v24, vb7, dv7, 0); \
+  v25 = vfma_lane_f16(v25, vb7, dv7, 1); \
+  v26 = vfma_lane_f16(v26, vb7, dv7, 2); \
+  v27 = vfma_lane_f16(v27, vb7, dv7, 3); \
+  dv7 = vld1_f16(a + 4 * 13);            \
+  vb7 = vld1_f16(b + 4 * 13);            \
+  v24 = vfma_lane_f16(v24, vb7, dv7, 0); \
+  v25 = vfma_lane_f16(v25, vb7, dv7, 1); \
+  v26 = vfma_lane_f16(v26, vb7, dv7, 2); \
+  v27 = vfma_lane_f16(v27, vb7, dv7, 3); \
+  dv7 = vld1_f16(a + 4 * 14);            \
+  vb7 = vld1_f16(b + 4 * 14);            \
+  v24 = vfma_lane_f16(v24, vb7, dv7, 0); \
+  v25 = vfma_lane_f16(v25, vb7, dv7, 1); \
+  v26 = vfma_lane_f16(v26, vb7, dv7, 2); \
+  v27 = vfma_lane_f16(v27, vb7, dv7, 3); \
+  dv7 = vld1_f16(a + 4 * 15);            \
+  vb7 = vld1_f16(b + 4 * 15);            \
+  v24 = vfma_lane_f16(v24, vb7, dv7, 0); \
+  v25 = vfma_lane_f16(v25, vb7, dv7, 1); \
+  v26 = vfma_lane_f16(v26, vb7, dv7, 2); \
+  v27 = vfma_lane_f16(v27, vb7, dv7, 3); \
+  l += 16;                                \
+  __builtin_prefetch(b + 64, 0, 3);      \
+  __builtin_prefetch(a + 64, 0, 3);      \
+  b += 4 * 16;                            \
+  a += 4 * 16;
+
+// 2. Partial sum 128 digits : Medium accuracy, medium latency
 #define KERNEL_4x4_ACC8()                \
   v24 = vdup_n_f16(0.F);                 \
   v25 = vdup_n_f16(0.F);                 \
@@ -91,6 +199,13 @@
   __builtin_prefetch(a + 4, 0, 3);       \
   b += 4 * 1;                            \
   a += 4 * 1;
+
+#define SAVE_KERNEL_4X4_F16_F32()                                       \
+  vst1q_f32(c, vaddq_f32(vld1q_f32(c), vcvt_f32_f16(v24)));             \
+  vst1q_f32(c + ldc, vaddq_f32(vld1q_f32(c + ldc), vcvt_f32_f16(v25))); \
+  vst1q_f32(c + 2 * ldc,                                                \
+            vaddq_f32(vld1q_f32(c + 2 * ldc), vcvt_f32_f16(v26)));      \
+  vst1q_f32(c + 3 * ldc, vaddq_f32(vld1q_f32(c + 3 * ldc), vcvt_f32_f16(v27)));
 
 /**
  * @brief hgemm 4x4 kernel sc = sa * sb
@@ -199,6 +314,8 @@ void hgemm_kernel_4x4(unsigned int M, unsigned int N, unsigned int K,
   __fp16 *a = sa, *b = sb;
   float *c = sc;
   unsigned int i, j, l;
+  unsigned int K16 = (K >> 4) << 4;
+  unsigned int K8 = (K >> 3) << 3;
   for (i = 0; i < M; i += VL_FP16_HALF) {
     for (j = 0; j < N; j += VL_FP16_HALF) {
       __builtin_prefetch(b, 0, 3);
@@ -207,16 +324,18 @@ void hgemm_kernel_4x4(unsigned int M, unsigned int N, unsigned int K,
       float16x4_t v24, v25, v26, v27;
       float16x4_t dv0, dv1, dv2, dv3, dv4, dv5, dv6, dv7;
       float16x4_t vb0, vb1, vb2, vb3, vb4, vb5, vb6, vb7;
-
-      for (l = 0; l < K;) {
+      l = 0;
+      for (; l < K16;) {
+        KERNEL_4x4_ACC16();
+        SAVE_KERNEL_4X4_F16_F32();
+      }
+      for (; l < K8;) {
         KERNEL_4x4_ACC8();
-
-        vst1q_f32(c, vaddq_f32(vld1q_f32(c), vcvt_f32_f16(v24)));
-        vst1q_f32(c + ldc, vaddq_f32(vld1q_f32(c + ldc), vcvt_f32_f16(v25)));
-        vst1q_f32(c + 2 * ldc,
-                  vaddq_f32(vld1q_f32(c + 2 * ldc), vcvt_f32_f16(v26)));
-        vst1q_f32(c + 3 * ldc,
-                  vaddq_f32(vld1q_f32(c + 3 * ldc), vcvt_f32_f16(v27)));
+        SAVE_KERNEL_4X4_F16_F32();
+      }
+      for (; l < K;) {
+        KERNEL_4x4_ACC1();
+        SAVE_KERNEL_4X4_F16_F32();
       }
 
       c += 4;