> Unittest output using Galaxy S23 with local commit

### Latency
> mean latency with TC = 100

| dim | KERNEL_8x16_ACC16 | KERNEL_8x16_ACC8 | cblas fp32 |
| --- | --- | --- | --- |
| 1024 | **23 ms** | 30 ms | 32 ms |
| 768 | **9 ms** | 12.8 ms | 13.6 ms |
| 256x1440x256 | **2054 mcrs** | 2664 mcrs  | 2701 mcrs |
| 256x256x1440 | **2359 mcrs** | 2965 mcrs  | 3104 mcrs |

### mse w.r.t. sgemm

| dim | KERNEL_8x16_ACC16 | KERNEL_8x16_ACC8 |
| --- | --- | --- |
| 1024 |  0.00608169 | 0.00226737 |
| 768 | 0.00310214 | 0.0017091 |
| 256x1440x256  | 0.0149112 | 0.00518965 |
| 256x256x1440 | 0.00119428 | 0.000306849 |

- Overall, this shows **150% boost-up** with f16-f32 w.r.t. cblas fp32
- Considering enlarged vector length from f32 to f16, and partial-accumulation, result above sounds reasonable.
- However, this code takes a little bit of accuracy loss for its cost. Should be checked once more with model output.

### + Trivial refactor:
- unified macro-style GEMM kernels.
- adaptive K-direction loops