@jijoongmoon 2024-05-10
what the ACC16 and ACC8 Means?