@skykongkong8 2024-05-10
> what the ACC16 and ACC8 Means?

It means the number of iterations of the tile of current gemm kernel.
Suppose GEMM(M, K, N, ... ).
Using axb gemm kernel with ACC'c' will therefore compute a*b*c digits with fp16, and accumulate with f32 M*K*N / (a*b*c) times.
That's why I wrote partial sum up to '2048 digits' with 8x16 ACC16 -> 8*16*16 = 2048